requirements.txt
# For faster and efficiant python packages installation
wheel
setuptools
# For Django app - https://www.djangoproject.com/
django
# For Django rest apis - https://www.django-rest-framework.org/
djangorestframework
# For reactjs - https://pypi.org/project/django-cors-headers/
django-cors-headers
# For image in python - https://pypi.org/project/Pillow/
pillow
# For Django JWT auth token - https://django-rest-framework-simplejwt.readthedocs.io/en/latest/
djangorestframework-simplejwt
# For PostGres database - https://pypi.org/project/psycopg2/
psycopg2
# Heroku - https://pypi.org/project/gunicorn/
gunicorn
# Heroku - http://whitenoise.evans.io/en/stable/
whitenoise
# Heroku - https://devcenter.heroku.com/articles/django-app-configuration
django-heroku
# Heroku PostGresql- https://devcenter.heroku.com/articles/heroku-postgresql#connecting-in-python
psycopg2-binary
dj-database-url
python-decouple